pyproject.toml
[tool.poetry]
name = "the-dozens-django"
version = "0.1.0"
description = "Conversion of the Dozens Roast API from Flask RestX to Django DRF framework"
authors = ["Terry Brooks"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
markdown = "^3.5"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
django-debug-toolbar = "^4.2.0"
logtail-python = "^0.2.7"
psycopg2-binary = "^2.9.9"
redis = {extras = ["hiredis"], version = "^5.0.1"}
django-mailer = "^2.3"
graphene-django = "^3.1.5"
uritemplate = "^4.1.1"
django-rest-swagger = "^2.2.0"
django-crispy-forms = "^2.1"
crispy-bootstrap5 = "^0.7"
django-cacheops = "^7.0.1"
ghapi = "^1.0.4"
gunicorn = "^21.2.0"
django-prometheus = "^2.3.1"
django-redis = "^5.4.0"

[tool.poetry.group.test.dependencies]
pytest-django = "^4.5.2"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pip = "^23.3"
install = "^1.3.5"
isort = "^5.12.0"
mypy = "^1.6.1"
pylint = "^3.0.2"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "test.settings"

[tool.djlint]
format_css=true
format_js=true
ignore="H021,T001,H008,H030,H031,H006"

[tool.djlint.css]
indent_size=5

[tool.djlint.js]
indent_size=5
